requirements.txt
streamlit
yfinance
pandas
plotly
firebase-admin
pyrebase4
joblib
requests
beautifulsoup4
tqdm
altair
streamlit-lottie
textblob
scikit-learn
tweepy
